Moved the parameter parsing to a separate module
